Corrections to hash code
The previous implementation was using a HashSet to store the hash
seeds. Thus for each new set a different order of seeds was generated
and thus a different hash code. The seed set has now be changed to a
collection and the Seed Factory now returns a LinkedHashSet that allows
us to keep the unique seeds whilst maintaining the order.

The starting molecule seed was also change from molsize % 389  to 389 %
molsize. The previous implementation was always returning the molecule
size and now returns a prime number (good for hashing).